third_party/WebKit/LayoutTests/jquery/attributes-expected.txt

CONSOLE WARNING: Attr.specified is deprecated. Its value is always true.
ALERT: 339 tests of 339 passed, 0 failed.
